requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
flask-socketio==5.3.6
eventlet>=0.36.1
PyJWT==2.8.0
requests==2.31.0
google-genai
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
numpy>=1.24.0
dnspython
setuptools
flask-sqlalchemy
flask-migrate
psycopg2-binary
fpdf2
